+++ docs/CHANGELOG.md
@@ -19,6 +19,7 @@ Version 1.6.1 (not yet released)
 - Add server-side CPS limit
 - Separate maxfps cvar for dedicated server and client
 - Clear console input field when Control+C is pressed
+- Do not add repeated commands to the console history
 
 Version 1.6.0 (released 2021-01-07)
 --------------------------------